convertdatetoutc in utils
